24 release action

+++ crates/gibberish_core/src/expected.rs
@@ -3,16 +3,16 @@ use crate::{err::Expected, lang::CompiledLang};
 #[repr(C)]
 #[derive(Clone, Copy)]
 pub struct ExpectedData {
-    kind: usize,
-    id: usize,
+    kind: u32,
+    id: u32,
 }
 
 impl From<ExpectedData> for Expected<CompiledLang> {
     fn from(value: ExpectedData) -> Self {
         match value.kind {
-            0 => Expected::Token(value.id as u32),
-            1 => Expected::Group(value.id as u32),
-            2 => Expected::Label(value.id as u32),
+            0 => Expected::Token(value.id),
+            1 => Expected::Group(value.id),
+            2 => Expected::Label(value.id),
             kind => panic!("Unsupported kind for expected {kind}"),
         }
     }
